refactor(ma): calendar -  wrapper: component, month-switcher: component and hook
