--- mother-frontend/micro-apps/calendar/src/ui/modules/month-switcher/month-switcher.style.scss
@@ -1,3 +1,0 @@
-.CalendarMonthSwitcher {
-  width: 80px;
-}
