--- mother-frontend/micro-apps/calendar/src/ui/modules/month-switcher/month-switcher.vue
@@ -1,19 +1,0 @@
-<template>
-  <!-- <UISelect
-    id="date-months"
-    class="CalendarMonthSwitcher"
-    :model-value="month"
-    :options="selectMonths"
-    @update:model-value="changeCalendarByMonth($event)"
-  /> -->
-</template>
-
-<script setup lang="ts">
-  // import UISelect from '@windows98/app/Bundles/UI/UISelectBundle/Components/UISelect.vue'
-
-  // import useMonthSwitcher from './use-month-switcher.composable'
-
-  // const { changeCalendarByMonth, month, selectMonths } = useMonthSwitcher()
-</script>
-
-<style lang="scss" scoped src="./month-switcher.style.scss" />
