+++ mother-frontend/micro-apps/calendar/src/ui/modules/month-switcher/use-month-switcher.hook.ts
@@ -1,0 +1,69 @@
+import { useState } from 'react';
+import { DSSelectOption } from '@windows98/design-system';
+
+const selectMonths: Array<DSSelectOption> = [
+  {
+    value: 0,
+    name: 'CalendarBundle.january',
+  },
+  {
+    value: 1,
+    name: 'CalendarBundle.february',
+  },
+  {
+    value: 2,
+    name: 'CalendarBundle.march',
+  },
+  {
+    value: 3,
+    name: 'CalendarBundle.april',
+  },
+  {
+    value: 4,
+    name: 'CalendarBundle.may',
+  },
+  {
+    value: 5,
+    name: 'CalendarBundle.june',
+  },
+  {
+    value: 6,
+    name: 'CalendarBundle.july',
+  },
+  {
+    value: 7,
+    name: 'CalendarBundle.august',
+  },
+  {
+    value: 8,
+    name: 'CalendarBundle.september',
+  },
+  {
+    value: 9,
+    name: 'CalendarBundle.october',
+  },
+  {
+    value: 10,
+    name: 'CalendarBundle.november',
+  },
+  {
+    value: 11,
+    name: 'CalendarBundle.december',
+  },
+]
+
+export default function useMonthSwitcher() {
+  const [month, setMonth] = useState<number>(Number(selectMonths[0].value));
+
+  const changeCalendarByMonth = (value: number): void => {
+    setMonth(value)
+    console.log(value)
+  }
+
+  return {
+    changeCalendarByMonth,
+    month,
+    selectMonths,
+  }
+}
+
