+++ mother-frontend/micro-apps/calendar/src/ui/modules/wrapper/wrapper.tsx
@@ -1,0 +1,9 @@
+import MonthSwitcher from '../month-switcher/month-switcher';
+
+export function Wrapper() {
+	return (
+		<div>
+			<MonthSwitcher />
+		</div>
+  )
+}
